Implemeted authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -63,7 +63,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (\Gate::allows('update-question', $question)) {
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, "Access denied");
     }
 
     /**
@@ -75,9 +78,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title', 'body'));
-        return redirect('questions')->with('success', "Your question has been updated.");
-
+         if (\Gate::allows('update-question', $question)) {
+            $question->update($request->only('title', 'body'));
+            return redirect('questions')->with('success', "Your question has been updated.");
+        }
+        abort(403, "Access denied");
     }
 
     /**
@@ -88,7 +93,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
-        return redirect('/questions')->with('success', "Your Question has been deleted.");
+         if (\Gate::allows('delete-question', $question)) {
+             $question->delete();
+             return redirect('/questions')->with('success', "Your Question has been deleted.");
+        }
+        abort(403, "Access denied");
+       
     }
 }
